refactor: move formatErrorMessage to common/message-handler and update imports
